Plot a smoothed density plot of the null distribution

Determines the rank of the mean absolute value of the standardized treatment effect using any placebo analyses that survive a user specified Cohen's D threshold ways to use the placebo distribution to calculate a mean value (e.g., mean p-value or another test statistic). This is a simple example of one such way.

PlotNullDistribution(
  CohensD = 999,
  StartTime = TreatmentBeginsAt,
  EndTime = nrow(SCUL.output$y.actual),
  width = 5,
  height = 2.76,
  AdjustmentParm = 1,
  BandwidthParm = 0.25,
  y_label = "",
  subtitle_label = "",
  x_label = "Distribution of standardized difference\n for placebo donor pool",
  title_label = "Cohen's-D restriction: None",
  rejection_label = "Rejection region\n for null hypothesis\n of no treatment effect",
  save.figure = FALSE,
  OutputFilePath = SCUL.input$OutputFilePath,
  TreatmentBeginsAt = SCUL.output$TreatmentBeginsAt
)

Arguments

CohensD
A real number greater than 0, indicating the Cohen's D threshold at which fit is determined to be "poor". The difference is in standard deviation units. Default is SCUL.input$CohensDThreshold.

StartTime
The begining time period for which the average pseduo treatment effect is calculated. T

EndTime
The end time period for which the average pseduo treatment effect is calculated. T

width
width of histogram. Default is 20

height
height of histogram. Default is 0.05

AdjustmentParm
Adjustment parameter for guassian histogram. Default is 3.

BandwidthParm
Bandwidth parameter for guassian histogram. Default is 3.

y_label
label for the y-axis

subtitle_label
Graph sub-title

x_label
label for the x-axis

title_label
Graph title

rejection_label
Label for the top of the red rejection region

save.figure
Boolean, set to TRUE if you want output saved as figure to OutputFilePath automatically. Default is FALSE

OutputFilePath
Output file path. Default is SCUL.input$OutputFilePath.

TreatmentBeginsAt
An integer indicating which row begins treatment. Default is SCUL.output$TreatmentBeginsAt.

Value

graph A denisty plot of the null distribution conditional on satisfactory fit